requirements.txt
requests==2.11.1
flask==0.11
flask-ask==0.7.8
gunicorn==19.6.0
psycopg2==2.6.2
flask-sqlalchemy==2.1
Flask-Migrate==1.8.0
